refactor(FE): `useFleeBattle` hook
